fix: adding error for delete league controller method

+++ server/controllers/league.controller.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { League, Group } = require("../models");
+const { League, Group, Season } = require("../models");
 const modelByPk = require("./utility");
 
 const LeagueController = function () {
@@ -98,9 +98,22 @@ const LeagueController = function () {
 
   var deleteLeague = async function (req, res, next) {
     try {
+      const leagueId = req.params["id"];
+      const hasSeasons = await Season.findAll({
+        where: {
+          league_id: leagueId,
+        },
+      });
+
+      if (hasSeasons.length > 0) {
+        return res.status(200).json({
+          error:
+            "league cannot be deleted if it has at least one season linked to it",
+        });
+      }
       let deletedLeague = await League.destroy({
         where: {
-          id: req.params["id"],
+          id: leagueId,
         },
       });
